[WIP] Reorganized error types into separate files.
